[chore/#60] 파일 중복으로 충돌 이슈 해결
